# Optimization: parallel level shuffle at sender side