… be/src/exec/pipeline/exchange/exchange_source_operator.cpp
… be/src/exec/pipeline/exchange/exchange_source_operator.h
… be/src/exec/pipeline/exchange/local_exchange.cpp
… be/src/exec/pipeline/exchange/local_exchange.h
… be/src/exec/pipeline/exchange/sink_buffer.cpp
… be/src/exec/pipeline/exchange/sink_buffer.h
… be/src/exec/pipeline/fragment_executor.cpp
… be/src/exec/pipeline/operator.h
… be/src/exec/pipeline/pipeline_builder.cpp
… be/src/exec/pipeline/pipeline_builder.h
… be/src/exec/vectorized/aggregate/aggregate_blocking_node.cpp
… be/src/exec/vectorized/hash_join_node.cpp
… be/src/exec/vectorized/hash_join_node.h
… be/src/exec/vectorized/hash_joiner.cpp
… be/src/exec/vectorized/hash_joiner.h
… be/src/runtime/data_stream_recvr.cc
… be/src/runtime/data_stream_recvr.h
… be/src/runtime/local_pass_through_buffer.cpp
… be/src/runtime/local_pass_through_buffer.h
… fe/fe-core/src/main/java/com/starrocks/planner/DataStreamSink.java
… fe/fe-core/src/main/java/com/starrocks/planner/ExchangeNode.java
… fe/fe-core/src/main/java/com/starrocks/planner/HashJoinNode.java
… fe/fe-core/src/main/java/com/starrocks/planner/PlanFragment.java
… fe/fe-core/src/main/java/com/starrocks/qe/Coordinator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/base/DistributionSpec.java
… fe/fe-core/src/main/java/com/starrocks/sql/plan/PlanFragmentBuilder.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/DecodeRewriteTest.java
… gensrc/proto/internal_service.proto
… gensrc/thrift/DataSinks.thrift
… gensrc/thrift/Partitions.thrift