# Background

For query like `select avg(*) from t0 group by c1`, some of the pipelines are listed as follows:
1. `<prev_op>` → `exchange_sink_op`
1. `exchange_source_op` → **`local_shuffle_exchange_sink_op`**
1. **`local_shuffle_exchange_source_op`** → `aggregate_blocking_sink_op`
1. `aggregate_blocking_source_op` → `<next_op>`
1. ...

`aggregate_blocking_op` can be parallelized if group by clause exists, we only need to ensure that the multiple inputs are orthogonal to each other.

So here is the solution (before optimization), chunks are first shuffled at machine(backend) level in the first pipeline, and chunks are second shuffled at parallel level in the third pipeline, **which performe hash operations for two times (It can be the boltneck in some cases).**


# How

How to optimize? The solution is kind of straightforward, we can doing shuffle only once both at machine and parallel level. And after optimization. The optimized pipelines are as follows:

1. `<prev_op>` → `exchange_sink_op`
1. `exchange_source_op` → `aggregate_blocking_sink_op`
1. `aggregate_blocking_source_op` → `<next_op>`
1. ...

**Thus, `local_shuffle_exchange_sink/source_op` is no longer needed in some cases.**

# Changes

1. message `PTransmitChunkParams` add optional field `shuffle_id`
1. `TExchangeNode` add optional field `partition_type`
1. `ExchangeSinkOperator` enable pipeline level shuffle when exchange's partition type is `TPartitionType::HASH_PARTITIONED`
1. each instance of `ExchangeSourceOperator` only process those data with `PTransmitChunkParams::shuffle_id` equals to `ExchangeSourceOperator::_driver_sequence` 
1. remove local shuffle operators before `AggregateBlockingSinkOperator`
1. remove local shuffle operators before `HashJoin{Build, Probe}Operator`
1. `ExchangeSourceOperator` support short circuit

# Test

**Settings:** 

1. fe(1), be(3)
2. 64G/16c
3. tpch_100g、ssb_100g
3. dop: 8

**Queries:**

```sql
-- Q1
select avg(L_QUANTITY) as avg from tpch_100g.lineitem group by L_PARTKEY order by avg limit 20;

-- Q2
SELECT SUM(c_custkey), SUM(char_length(c_name)), SUM(char_length(c_address)),SUM(char_length(c_city)), SUM(char_length(c_nation)), SUM(char_length(c_region)), SUM(char_length(c_phone)), SUM(char_length(c_mktsegment)) FROM ssb_100g.lineorder INNER JOIN ssb_100g.customer ON lo_custkey=c_custkey;
```

**Results:**

| Query | Non-pipeline | Before Optimization | After Optimization |
|:--|:--|:--|:--|
| Q1 | 4.70s | 5.04s | 4.31s |
| Q2 | 9.08s | 10.88s | 9.09s |


# Todo

* MultiCastDataStreamSink
* The judgement at receiver side(`HashJoinNode::decompose_to_pipeline`) on whether need local shuffle may not be optimal
* Other operators may also remove local shuffle