@ZiheLiu 2022-01-17
Maybe all the operators, whose preceding is `ExchangeSource`, can eliminate the local shuffle exchange? For example, ExceptNode, IntersectNode, SortNode in analytic.
Could we determine this in `maybe_interpolate_local_shuffle_exchange()` instead of `HashJoinNode::decompose_to_pipeline()` and `AggregateBlockingNode::decompose_to_pipeline().